Support multi-loop unroll jam using clast
The loops that need to be unroll jammed are marked in the clast by setting
`clast_for->unroll` field to CLAST_UNROLL_JAM. The marked loop is then unroll
jammed provided the loop nest is rectangular. The implementation also supports
multi-loop unroll jam, i.e, for a given statement, multiple loops enclosing the
statement can be unroll jammed. The routine also supports unroll jamming of
imperfect nested loops. The routine does not check for correctness of unroll
jam (permutability of loop nest). It is the responsibility of the caller to
mark the correct loops for unroll jam.

Routines added in this commit include
        - clast_unroll_jam which is the library interface routine for
          unroll jamming the AST.
        - Routines to deep copy AST. This aids generation of clean-up code
          after unroll jam.
        - Routines to check rectangularity of the loop nest. Currently, it
          checks whether the iterator of the loop being unroll jammed appears
          in the bounds of inner loops or guards of the statements enclosed.